requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0
openai>=1.0.0
packaging>=21.0
pyyaml>=6.0
fpdf2>=2.7.4
python-dotenv>=1.0.0
